base.html extendable and moved site_config to config.py
made base.html extendable and switched the routes over to index.html which extends it.

I made siteConfig.py which contains the logic to constructe the SiteConfig Named tuple. It is imported by Config.py which uses it to define the site_config variable. __init___.py then imports this to be used to define the sitewide variables. Routes then import this object as the config kwarg.

To Add/remove/change SiteConfig. Update siteConfig.py, config.py, and then the templates.